Move utils under tests
