+++ package.json
@@ -5,12 +5,12 @@
   "main": "index.js",
   "scripts": {
     "start": "npx ts-node src/main/index.ts",
-    "test": "./src/utils/init && NODE_ENV=test ./node_modules/mocha/bin/mocha  --timeout 10000 --require ts-node/register ./tests/*.spec.ts || true && ./src/utils/destroy || true",
-    "test-on-pre-existing-cluster": "./src/utils/init-assume-cluster-exists && NODE_ENV=test ./node_modules/mocha/bin/mocha  --timeout 10000 --require ts-node/register ./tests/*.spec.ts",
+    "test": "./tests/utils/init && NODE_ENV=test ./node_modules/mocha/bin/mocha  --timeout 10000 --require ts-node/register ./tests/*.spec.ts || true && ./tests/utils/destroy || true",
+    "test-on-pre-existing-cluster": "./tests/utils/init-assume-cluster-exists && NODE_ENV=test ./node_modules/mocha/bin/mocha  --timeout 10000 --require ts-node/register ./tests/*.spec.ts",
     "eslint": "eslint src/**/*.ts src/**/*.js tests/**/*.ts",
-    "coverage": "./src/utils/init && NODE_ENV=test nyc ./node_modules/mocha/bin/mocha  --timeout 10000 --require ts-node/register ./tests/*.spec.ts || true && ./src/utils/destroy || true",
-    "e2e": "./src/utils/init && NODE_ENV=test ./node_modules/mocha/bin/mocha  --timeout 20000 --require ts-node/register ./tests/e2e/*.spec.ts || true && ./src/utils/destroy || true",
-    "e2e-on-pre-existing-cluster": "./src/utils/init-assume-cluster-exists && NODE_ENV=test ./node_modules/mocha/bin/mocha  --timeout 20000 --require ts-node/register ./tests/e2e/*.spec.ts",
+    "coverage": "./tests/utils/init && NODE_ENV=test nyc ./node_modules/mocha/bin/mocha  --timeout 10000 --require ts-node/register ./tests/*.spec.ts || true && ./tests/utils/destroy || true",
+    "e2e": "./tests/utils/init && NODE_ENV=test ./node_modules/mocha/bin/mocha  --timeout 20000 --require ts-node/register ./tests/e2e/*.spec.ts || true && ./tests/utils/destroy || true",
+    "e2e-on-pre-existing-cluster": "./tests/utils/init-assume-cluster-exists && NODE_ENV=test ./node_modules/mocha/bin/mocha  --timeout 20000 --require ts-node/register ./tests/e2e/*.spec.ts",
     "build-docker-image": "docker build . -t medicmobile/upgrade-service:0.26",
     "build-docker-image-latest": "docker build . -t medicmobile/upgrade-service"
   },
