+++ tests/utils/init-assume-cluster-exists
